requirements.txt
requests>=2.20.0
python-dotenv>=0.10.0
python-docx>=0.8.11
pandas>=1.3.0
openpyxl>=3.0.0

flask==2.3.3
flask-cors==4.0.0
requests==2.31.0
python-dotenv==1.0.0
werkzeug==2.3.7
asyncpg>=0.27.0